fix: a dead covariate must set no scale on every interpreter, not just some
Two CI failures, both real, both from a batch of eleven commits that had never
been through CI. Neither reproduces on this machine, which is the point.

1. benchmark_gamma reported a confounding strength of 4.9e15 for a covariate
   that carries no signal at all (test 3.11, 3.13; passes on 3.12, 3.14).

   The gap between the full and reduced propensity comes from two INDEPENDENT
   fits, so dropping a column that explains nothing leaves a residue at rounding
   scale rather than exactly zero. exp() maps that 2.2e-16 to 1 + 2.2e-16, the
   `strongest <= 1.0` guard is an exact comparison and lets it through, and
   log() in the divisor turns the residue back into 2.2e-16 -- amplifying pure
   rounding noise by sixteen orders of magnitude. Whether it fires depends on
   which way exp() rounds, hence the split across interpreters: numpy 2.4.6
   resolves below Python 3.12 and 2.5.1 above it.

   Fixed in the producer. The quantile gap is snapped to zero below the
   backward-stability scale of the matvec that produced the logits,
   sqrt(n) * eps * |logit|, so a null covariate yields exactly 1.0 and the
   existing guard returns inf as it was written to. Measured here: the residue
   is 2.8e-17 locally and 2.2e-16 on the runners against a floor of 5.0e-15,
   while a deliberately weak but real covariate (beta = 0.05) sits at 1.1e-1 --
   thirteen orders above the floor, which a new test fences off.

   The assertion moves from approx(1.0) to exactly 1.0: approx is what let one
   ulp through, and the invariant really is exact.

2. ty rejected kronecker_spectrum's return on 3.11 only. numpy's stub for
   svd(compute_uv=False) widened to floating[Any] before 2.5, so the signature's
   float64 promise was resting on the stub rather than on the code. Now made
   true of the value with astype(copy=False), which is a no-op on the normal
   path.

Verification: ruff, ruff format, ty clean; full suite 609 passed, 2 skipped.

+++ src/chc/network_causal.py
@@ -198,7 +198,10 @@ def kronecker_spectrum(
         .transpose(0, 2, 1, 3)
         .reshape(n_units * n_units, n_times * n_times)
     )
-    return np.linalg.svd(rearranged, compute_uv=False)
+    # astype, not a cast: the signature promises float64 and this makes that true of the VALUE.
+    # numpy's own stub for compute_uv=False widened to floating[Any] before 2.5, so relying on the
+    # stub made the promise hold on one interpreter and not another.
+    return np.linalg.svd(rearranged, compute_uv=False).astype(np.float64, copy=False)
 
 
 def ar1_innovations(
